Major refactor to support cite styles like org-cite
See issue #9.

This fully supports abbreviations and full style names as far as I know.

I feel it will probably bite me later. This code works at the limit of my
cleverness. In my experience, debugging is twice as hard as writing the code, so
it is probable in the future I will not be clever enough to fix bugs in it...